// types/restaurant-old.d.ts
type RestaurentMenu = {
  /**店铺名 */
  title: string;
  /**描述*/
  description: string;
  /**描述html */
  description_html: string;
  /**优惠价格 */
  price: number;
  /** 原始价格，可能为空，没有表示没做过特价活动*/
  base_price: number;
  /**会员价格*/
  member_price: number;
  /**口味*/
  tags: string[];
  /**商品图片 */
  img_url: string;
  /**商品大图 */
  img_url_big: string;
  /**营业状态*/
  business_status: string;
  /**售卖地区*/
  zone_id: string;
  /**商品id*/
  product_id: number;
  /**库存 */
  quantity: number;
  /**已售数量*/
  quantity_sale: number;
  /** 最大可加购数量 */
  max_order_quantity: number;
  /** ui 扩展属性*/
  cartQty: number;
  vender_id: number;
  /** 超过5，是就打折扣标签,如果不是按折扣设的特价就没有，如果没有设优惠价也没有，默认是null，null就不打标签 */
  discount_percentage: number;
  /**店铺名 */
  vender_name?: string;
  /**价格类型 */
  price_type?: string;
  source?: string;
};
// ...
export type RestaurentMenuProps = Partial<RestaurentMenu>;

export type filterTagsProps = {
  label: string;
  key: number;
};

export type CuisineTypes = {
  img_url: string;
  key: number;
  label: string;
};

export type categoryProps = Partial<{
  category_id: number;
  category_name: string;
  dishes: RestaurentMenuProps[];
}>;